refactor: Anomaly and Segmentation evaluations now extend Evaluation task
