+++ quadra/tasks/base.py
@@ -1,4 +1,6 @@
+import json
 import os
+from pathlib import Path
 from typing import Any, Dict, Generic, List, Optional, TypeVar, Union
 
 import hydra
@@ -290,46 +292,73 @@ class PlaceholderTask(Task):
         log.info("If you are reading this, it means that library is installed correctly!")
 
 
-class Evaluation(Task):
+class Evaluation(Generic[DataModuleT], Task[DataModuleT]):
     """Base Evaluation Task with deployment models.
 
     Args:
         config: The experiment configuration
         model_path: The model path.
-        report_folder: The report folder. Defaults to None.
+        device: Device to use for evaluation. If None, the device is automatically determined.
 
-    Raises:
-        ValueError: If the experiment path is not provided
     """
 
     def __init__(
         self,
         config: DictConfig,
         model_path: str,
-        report_folder: Optional[str] = None,
+        device: Optional[str] = None,
     ):
         super().__init__(config=config)
+
+        if device is None:
+            self.device = utils.get_device()
+        else:
+            self.device = device
+
         self.config = config
-        self.metadata = {"report_files": []}
+        self.model_data: Dict[str, Any]
         self.model_path = model_path
         self.device = utils.get_device()
-        self.report_folder = report_folder
         self._deployment_model: RecursiveScriptModule
         self.deployment_model_type: str
-        if self.report_folder is None:
-            log.warning("Report folder is not provided, using default report folder")
-            self.report_folder = "report"
+        self.model_info_filename = "model.json"
+        self.report_path = ""
+        self.metadata = {"report_files": []}
 
     @property
-    def deployment_model(self) -> RecursiveScriptModule:
-        """RecursiveScriptModule: The deployment model."""
+    def deployment_model(self):
+        """Deployment model."""
         return self._deployment_model
 
     @deployment_model.setter
-    def deployment_model(self, model: RecursiveScriptModule) -> None:
-        """RecursiveScriptModule: The deployment model."""
-        self._deployment_model = model
+    def deployment_model(self, model_path: str):
+        """Set the deployment model."""
+        self._deployment_model, self.deployment_model_type = import_deployment_model(  # type: ignore[assignment]
+            model_path, self.device
+        )
 
     def prepare(self) -> None:
         """Prepare the evaluation."""
-        self.deployment_model, self.deployment_model_type = import_deployment_model(self.model_path, self.device)
+        with open(os.path.join(Path(self.model_path).parent, self.model_info_filename)) as f:
+            self.model_data = json.load(f)
+
+        if not isinstance(self.model_data, dict):
+            raise ValueError("Model info file is not a valid json")
+
+        if self.model_data["input_size"][0] != self.config.transforms.input_height:
+            log.warning(
+                f"Input height of the model ({self.model_data['input_size'][0]}) is different from the one specified "
+                + f"in the config ({self.config.transforms.input_height}). Fixing the config."
+            )
+            self.config.transforms.input_height = self.model_data["input_size"][0]
+
+        if self.model_data["input_size"][1] != self.config.transforms.input_width:
+            log.warning(
+                f"Input width of the model ({self.model_data['input_size'][1]}) is different from the one specified "
+                + f"in the config ({self.config.transforms.input_width}). Fixing the config."
+            )
+            self.config.transforms.input_width = self.model_data["input_size"][1]
+
+        self.deployment_model = self.model_path
+
+        super().prepare()
