+++ quadra/tasks/segmentation.py
@@ -183,7 +183,7 @@ class Segmentation(Generic[SegmentationDataModuleT], LightningTask[SegmentationD
                     for task in eval_tasks:
                         for file in task.metadata["report_files"]:
                             mflow_logger.experiment.log_artifact(
-                                run_id=mflow_logger.run_id, local_path=file, artifact_path=task.report_folder
+                                run_id=mflow_logger.run_id, local_path=file, artifact_path=task.report_path
                             )
 
                 if tensorboard_logger is not None and self.config.core.get("upload_artifacts"):
@@ -206,13 +206,12 @@ class Segmentation(Generic[SegmentationDataModuleT], LightningTask[SegmentationD
                                 utils.upload_file_tensorboard(file, tensorboard_logger)
 
 
-class SegmentationEvaluation(Evaluation):
+class SegmentationEvaluation(Evaluation[SegmentationDataModuleT]):
     """Segmentation Evaluation Task with deployment models.
 
     Args:
         config: The experiment configuration
         model_path: The experiment path.
-        report_folder: The report folder. Defaults to None.
 
     Raises:
         ValueError: If the model path is not provided
@@ -222,14 +221,9 @@ class SegmentationEvaluation(Evaluation):
         self,
         config: DictConfig,
         model_path: str,
-        report_folder: Optional[str] = None,
     ):
-        super().__init__(config=config, model_path=model_path, report_folder=report_folder)
+        super().__init__(config=config, model_path=model_path)
         self.config = config
-        self.metadata = {"report_files": []}
-
-        # TODO: It's not possible to specify the device from outside!!
-        self.device = utils.get_device(config.trainer.accelerator != "cpu")
 
     def save_config(self) -> None:
         """Skip saving the config."""
@@ -269,17 +263,11 @@ class SegmentationAnalysisEvaluation(SegmentationEvaluation):
     Args:
         config: The experiment configuration
         model_path: The model path.
-        report_folder: The report folder. Defaults to "analysis_report".
     """
 
-    def __init__(self, config: DictConfig, model_path: str, report_folder: str = "analysis_report"):
-        super().__init__(config=config, model_path=model_path, report_folder=report_folder)
+    def __init__(self, config: DictConfig, model_path: str):
+        super().__init__(config=config, model_path=model_path)
         self.test_output: Dict[str, Any] = {}
-
-    def prepare(self) -> None:
-        """Prepare the experiment."""
-        super().prepare()
-        self.datamodule = self.config.datamodule
 
     def train(self) -> None:
         """Skip training."""
